First step towards moving addesses from the attributes to own elements (-> feature whish #3133031). Also inital step towards feature whish #3134595 (Change data types to transfer functions)
As this will break all existing config files and as this is only the first step, that converted the info widget, there's a new config file (visu_config_neu.xml) that can be called by adding "?config=neu" to the visu URL. This file contains the new config syntax and will be extended by each widget that gets transformed.

NOTE: this will also break the editor!

So if you aren't perfectly sure about what you are doing, DO NOT update to the latest SVN version yet!